India Demo - BE/requirements.txt
amqp==5.0.9
aniso8601==9.0.1
APScheduler==3.9.1
bcrypt==3.2.0
bidict==0.21.4
billiard==3.6.4.0
blis==0.7.7
boto3==1.20.26
botocore==1.23.26
catalogue==2.0.7
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
chardet==5.0.0
charset-normalizer==2.0.9
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudpickle==2.0.0
cryptography==36.0.1
cymem==2.0.6
dask==2021.12.0
Deprecated==1.2.13
elastic-transport==8.1.2
elasticsearch==8.1.2
et-xmlfile==1.1.0
Flask==2.0.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
Flask-SocketIO==5.1.1
fsspec==2022.1.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
kombu==5.2.3
langcodes==3.3.0
locket==0.2.1
MarkupSafe==2.0.1
memory-profiler==0.60.0
murmurhash==1.0.7
mypy==0.930
mypy-extensions==0.4.3
mysql-connector-python==8.0.28
nameparser==1.1.1
numpy==1.21.5
openpyxl==3.0.9
packaging==21.3
pandas
paramiko==2.9.2
partd==1.2.0
pathy==0.6.1
pgeocode==0.3.0
preshed==3.0.6
prompt-toolkit==3.0.26
protobuf==3.19.4
psutil==5.9.0
psycopg2-binary==2.9.3
pycparser==2.21
pycryptodome==3.14.1
pycryptodome-test-vectors==1.0.7
pydantic==1.8.2
PyJWT==2.3.0
pymodm==0.4.3
pymongo==3.12.3
PyMySQL==1.0.2
PyNaCl==1.5.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-engineio==4.3.1
python-http-client==3.3.7
python-socketio==5.5.1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pyxlsb==1.0.9
PyYAML==6.0
redis==4.1.0
requests==2.28.1
s3transfer==0.5.0
sendgrid==6.9.7
simple-websocket==0.5.0
six==1.16.0
smart-open==5.2.1
spacy==3.3.0
spacy-legacy==3.0.9
spacy-loggers==1.0.2
SQLAlchemy==1.4.29
srsly==2.4.3
starkbank-ecdsa==2.0.3
thinc==8.0.16
tomli==2.0.0
toolz==0.11.2
tqdm==4.64.0
typer==0.4.1
typing==3.7.4.3
typing_extensions==4.0.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.7
vine==5.0.0
wasabi==0.9.1
wcwidth==0.2.5
Werkzeug==2.0.2
wrapt==1.13.3
wsproto==1.0.0
xlrd==2.0.1
xlsx2csv==0.7.8
